atomics: Remove requirement of 64 bit atomics
Remove two of the three  instances of components requiring
64 bit atomics, even on 32 bit systems.  The SM OSC component
also uses 64 bit atomics, but is a more complicated fix that
will follow this one.  Currently, no one is testing on
platforms that don't provide 64 bit atomics (even in 32 bit
mode), but with the removal of the non-inline assembly for
IA32, the older compilers on Absoft's test systems now
result in no practical way to call cmpxchg8 in 32 bit mode.
At that point, these failures started popping up.

+++ ompi/mca/osc/monitoring/osc_monitoring_module.h
@@ -2,6 +2,8 @@
  * Copyright (c) 2016-2017 Inria.  All rights reserved.
  * Copyright (c) 2017      Research Organization for Information Science
  *                         and Technology (RIST). All rights reserved.
+ * Copyright (c) 2017      Amazon.com, Inc. or its affiliates.  All Rights
+ *                         reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -34,7 +36,7 @@
 
 /* Define the ompi_osc_monitoring_module_## template ##_init_done variable */
 #define OMPI_OSC_MONITORING_MODULE_INIT_GENERATE(template)	\
-    static int64_t OMPI_OSC_MONITORING_MODULE_INIT(template)
+    static int32_t OMPI_OSC_MONITORING_MODULE_INIT(template)
 
 /* Define and set the ompi_osc_monitoring_## template ##_template
  * variable. The functions recorded here are linked to the original
